+++ src/getDepartments/index.ts
@@ -1,6 +1,6 @@
 import { departmentsMap } from "../internal/indexes";
 
-export function getDepartments() {
+export function getDepartments(): { code: string; name: string }[] {
   return Array.from(departmentsMap.entries()).map(([code, name]) => ({
     code,
     name,