+++ src/getUbigeoData/index.ts
@@ -1,7 +1,8 @@
 import { ubigeoMap } from "../internal/indexes";
 import { safeUbigeo6 } from "../internal/safeCode";
+import { UbigeoData } from "../types";
 
-export function getUbigeoData(ubigeo: unknown) {
+export function getUbigeoData(ubigeo: unknown): UbigeoData | null {
   const code = safeUbigeo6(ubigeo);
   if (!code) return null;